Created new system for saving to be able to run different versions of the code so we don't overwrite things by accident. Code now also has functionality to deal with if we want to run communities with a different total number of species (i.e. instead of 50 species per community, we run 100). Only half tested.
Fixed a few issues with reading in file paths for community data - was not reading in properly because was not loaded as a character.

Copied Susan's changes to the adaptive sampling script to the function to source.

Got rid of the specifying place to save raster in line with the new code for version controlling.