Fixed some problems with the adaptive sample function.
pars file was read in with file paths as 'factors' rather than characters - have changed code to recognise this.

When eff_layer was set as NULL, the column names retained the name of the variable used from the env_extent rasterstack. I have added renaming columns to deal with this.

Added code to stop the function if the method is coverage and no background sampling layer has been specified.

Finally, have sorted error out where very rare species did not get any new sightings in the adaptive sampling process. This threw an error but have added an if statement to deal with this.

Very minor edits elsewhere.