Added some code for testing.
Fixed a mistake with the coverage method which unvisited cells in the butterfly effort layer. Changed it to creating an observation data frame which is then used to find the unvisited locations. This is code stolen from the uncertainty + recs method so should be okay. Have tested but not submitted.